move: 정확한 표현을 위해 패키지 이름 변경
related to: #16
